sesame/requirements.txt
conan>=1.9.0
conan-package-tools>=0.20.0
bincrafters-package-tools>=0.17.5
